-Eye Glows are now a Character Option! Choose from None, Default, and Death Knight! Known Bug: Currently defaults to None on character load. -LWO Exporter:  -New M2 Exporter now serving as main LWO M2 Exporter.  -M2 Exporter now supports Luminous surfaces.  -Added Options to disable Camera & Bone exporting. (Not Yet Functional)  -(Dev Work) Doodads Sets as Layers exporting is 95% done. There's currently a bug that malforms some models. (The DwarvenBones01.mdx from altarofstorms, for example. Oddly, enough, it works perfectly when just exporting the DwarvenBones01.m2 model.)  -(Dev Work) Major reworking of various structures for better support.  -(Dev Work) Started work on new Scene-writing functions. -Dev Work:  -64-bit GLEW libs renamed to glew64.lib.  -Increased 64-bit compatibility.  -Added Matrix to Quaternion function.  -M2's animGeometry, animTextures and animBones boolans are now public.

+++ trunk/src/animcontrol.cpp
@@ -311,7 +311,7 @@ void AnimControl::UpdateWMO(WMO *w, int group)
 	wmoList->Show(TRUE);
 
 	// get wmo name or current wmogroup name/descr
-	if (group>=-1 && group<g_selWMO->nGroups) {
+	if (group>=-1 && group<(int)g_selWMO->nGroups) {
 		wxString label = w->name.AfterLast('\\');
 		if (group>=0) {
 			label += _T(" - ") + g_selWMO->groups[group].name;
